Correct random number call for windows to build tests

+++ include/carve/win32.h
@@ -16,7 +16,7 @@ inline void srandom(unsigned long input) {
   srand(input);
 }
 
-inline long random() {
+inline int random() {
   return rand();
 }
